Anal retentive code re-organization. Try and move the primitive computational geometry functions into lwalgorithm.c
git-svn-id: http://svn.osgeo.org/postgis/trunk@10356 b70326c6-7e19-0410-871a-916f4a2858ee

+++ liblwgeom/cunit/cu_algorithm.c
@@ -61,7 +61,7 @@ static int clean_cg_suite(void)
 */
 static void test_lw_segment_side(void)
 {
-	double rv = 0.0;
+	int rv = 0;
 	POINT2D p1, p2, q;
 
 	/* Vertical line at x=0 */
@@ -75,19 +75,19 @@ static void test_lw_segment_side(void)
 	q.y = 1.5;
 	rv = lw_segment_side(&p1, &p2, &q);
 	//printf("left %g\n",rv);
-	CU_ASSERT(rv < 0.0);
+	CU_ASSERT(rv < 0);
 
 	/* On the right */
 	q.x = 2.0;
 	rv = lw_segment_side(&p1, &p2, &q);
 	//printf("right %g\n",rv);
-	CU_ASSERT(rv > 0.0);
+	CU_ASSERT(rv > 0);
 
 	/* On the line */
 	q.x = 0.0;
 	rv = lw_segment_side(&p1, &p2, &q);
 	//printf("on line %g\n",rv);
-	CU_ASSERT_EQUAL(rv, 0.0);
+	CU_ASSERT_EQUAL(rv, 0);
 
 }
